Cancel merge_node operations and subsequent operations
When the user deletes an expanded selection, Slate attempts to merge the first and last nodes. At this point, we don't want to merge anything. Since the merge_node operations are the last to occur (prior to operations initiated by normalizers), we should just cancel the merge operation and anything after it. (If we don't cancel operations after the merge, we will be out of sync with what slate thinks happened- because we don't apply the merge operations.)